requirement.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
future==0.18.2
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==2.3.0
pymemcache==3.5.2
pyparsing==3.0.7
pytz==2021.3
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
